+++ src/Truck.js
@@ -1,8 +1,8 @@
 export default class Truck {
   constructor(length, width, height) {
-  	this.length = length;
-  	this.width = width;
-  	this.height = height;
+    this.length = length;
+    this.width = width;
+    this.height = height;
     this.items = [];
   }
 }
